// app/api/models/models.ts
export enum UserRoles {
	staff = 'staff',
	admin = 'admin',
	teller = 'teller',
	branchManager = 'branch_mgr',
	dftInputer = 'dft_input',
	dftAuthorizer = 'dft_auth',
	tropsInputer = 'treasuryops_inp',
	tropsAuthorizer = 'trops_auth',
	zonalHead = 'zHead',
	groupHead = 'gHead',
	ftInputer = 'ft_input',
	ftAuth = 'ft_auth',
	cco = 'cco_branch',
	ccoBranchAuth = 'cco_branch_auth',
	reviewer = 'bvndesk_input',
	chqMgmt = 'chqmgmt_input',
	// ===========new
	zonalHeadAut = 'zonal_head_aut',
	bvndeskAuth = 'bvndesk_auth',
	cdmsInput = 'cdms_input',
	cdmsAuth = 'cdms_auth',
	complianceInp = 'compliance_inp',
	complianceAuth = 'compliance_auth',
	intauditInput = 'intaudit_input',
	bvnReviewer = 'bvn_reviewer',
	ninValidator = 'nin_validator',
	headOps = 'head_of_ops',
	assistheadOps = 'assistant_head_of_ops',
	legal = 'legal',
	rManagerAuth = 'rel_mgr_auth',
	contactCenterAuth = 'contact_center_auth',
	contactCenterInputter = 'contact_center_inputter',
	bvnRequestGroup = 'bvn_request_group',
	bvnCCORequestGroup = 'bvn_cco_request_group',
	bvnAuthorizerGroup = 'bvn_authorizer_group',
	bvnCompleteRequestGroup = 'bvn_complete_request_group',
	ninCompGroup = 'nin_comp_group',
}